+++ R/ggpredict.R
@@ -637,7 +637,7 @@ ggpredict_helper <- function(model,
   if (model_class == "coxph" && type == "surv") model_info$is_binomial <- TRUE
 
   # get model frame
-  model_frame <- insight::get_data(model, source = "frame")
+  model_frame <- .get_model_data(model)
 
   # expand model frame to data grid of unique combinations
   data_grid <- .data_grid(