+++ tests/testthat/test-brms-trial.R
@@ -7,10 +7,11 @@ if (.runThisTest &&
     requiet("ggeffects") &&
     requiet("insight")
   )) {
-  m1 <- insight::download_model("brms_mixed_6")
+  m1 <- suppressWarnings(insight::download_model("brms_mixed_6"))
   m2 <- insight::download_model("brms_mv_4")
   m3 <- insight::download_model("brms_2")
 
+  skip_if(is.null(m1) || is.null(m2) || is.null(m3))
   test_that("ggpredict, brms-trial", {
     ggpredict(m1, c("Base", "Trt"))
     ggpredict(m2, "Species")