+++ tests/testthat/test-scoping.R
@@ -1,18 +1,18 @@
 if (suppressWarnings(requiet("testthat") && requiet("ggeffects") && requiet("insight"))) {
 
   test_that("ggpredict, scoping", {
-    data(efc)
+    data(efc, package = "ggeffects")
 
     fn1 <- function(data) {
       fit <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = data)
       v <- c(20, 50, 70)
-      ggpredict(fit, terms = "c12hour [20, 50, 70]")
+      ggpredict(fit, terms = "c12hour [20, 50, 70]", verbose = FALSE)
     }
 
     fn2 <- function(data) {
       fit <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = data)
       v <- c(20, 50, 70)
-      ggpredict(fit, terms = "c12hour [v]")
+      ggpredict(fit, terms = "c12hour [v]", verbose = FALSE)
     }
 
     out1 <- capture.output(print(fn1(efc)))
@@ -21,10 +21,10 @@ if (suppressWarnings(requiet("testthat") && requiet("ggeffects") && requiet("ins
   })
 
   test_that("ggpredict, scoping-2", {
-    data(efc)
+    data(efc, package = "ggeffects")
     fn3 <- function(data, v) {
       fit <- lm(barthtot ~ c12hour + c172code, data = data)
-      ggpredict(fit, terms = "c12hour [v]")
+      ggpredict(fit, terms = "c12hour [v]", verbose = FALSE)
     }
     v <- c(5, 6, 7)
     a <- c(0.3, 0.5, 0.7)